+++ src/main/java/ok/dht/test/monakhov/database/Config.java
@@ -2,8 +2,4 @@ package ok.dht.test.monakhov.database;
 
 import java.nio.file.Path;
 
-public record Config(
-    Path basePath,
-    long flushThresholdBytes)
-{
-}
+public record Config(Path basePath, long flushThresholdBytes) {}
