+++ src/main/java/ok/dht/test/monakhov/database/EntryKeyComparator.java
@@ -4,7 +4,7 @@ import jdk.incubator.foreign.MemorySegment;
 
 import java.util.Comparator;
 
-public class EntryKeyComparator implements Comparator<Entry<MemorySegment>> {
+public final class EntryKeyComparator implements Comparator<Entry<MemorySegment>> {
 
     public static final Comparator<Entry<MemorySegment>> INSTANCE = new EntryKeyComparator();
