+++ src/main/java/ok/dht/test/monakhov/database/MemorySegmentComparator.java
@@ -5,7 +5,7 @@ import jdk.incubator.foreign.MemorySegment;
 
 import java.util.Comparator;
 
-public class MemorySegmentComparator implements Comparator<MemorySegment> {
+public final class MemorySegmentComparator implements Comparator<MemorySegment> {
 
     public static final Comparator<MemorySegment> INSTANCE = new MemorySegmentComparator();
 
@@ -24,9 +24,7 @@ public class MemorySegmentComparator implements Comparator<MemorySegment> {
         if (firstMismatch == m2.byteSize()) {
             return 1;
         }
-        return Byte.compareUnsigned(
-            MemoryAccess.getByteAtOffset(m1, firstMismatch),
-            MemoryAccess.getByteAtOffset(m2, firstMismatch)
-        );
+        return Byte.compareUnsigned(MemoryAccess.getByteAtOffset(m1, firstMismatch),
+            MemoryAccess.getByteAtOffset(m2, firstMismatch));
     }
 }
