+++ src/main/java/ok/dht/test/monakhov/database/Storage.java
@@ -9,13 +9,17 @@ import java.io.IOException;
 import java.lang.ref.Cleaner;
 import java.nio.channels.FileChannel;
 import java.nio.file.Files;
-import java.nio.file.NoSuchFileException;
 import java.nio.file.Path;
 import java.nio.file.StandardCopyOption;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Iterator;
+import java.util.List;
 import java.util.concurrent.ThreadFactory;
+
+import static ok.dht.test.monakhov.database.StorageUtils.getSize;
+import static ok.dht.test.monakhov.database.StorageUtils.mapForRead;
+import static ok.dht.test.monakhov.database.StorageUtils.writeRecord;
 
 class Storage implements Closeable {
 
@@ -41,6 +45,10 @@ class Storage implements Closeable {
     private static final String FILE_EXT_TMP = ".tmp";
     private static final String COMPACTED_FILE = FILE_NAME + "_compacted_" + FILE_EXT;
 
+    private final ResourceScope scope;
+    private final List<MemorySegment> sstables;
+    private final boolean hasTombstones;
+
     static Storage load(Config config) throws IOException {
         Path basePath = config.basePath();
         Path compactedFile = config.basePath().resolve(COMPACTED_FILE);
@@ -51,14 +59,10 @@ class Storage implements Closeable {
         ArrayList<MemorySegment> sstables = new ArrayList<>();
         ResourceScope scope = ResourceScope.newSharedScope(CLEANER);
 
-        // FIXME check existing files
-        for (int i = 0; ; i++) {
-            Path nextFile = basePath.resolve(FILE_NAME + i + FILE_EXT);
-            try {
-                sstables.add(mapForRead(scope, nextFile));
-            } catch (NoSuchFileException e) {
-                break;
-            }
+        Path nextFile = basePath.resolve(FILE_NAME + 0 + FILE_EXT);
+        for (int i = 1; Files.exists(nextFile); i++) {
+            sstables.add(mapForRead(scope, nextFile));
+            nextFile = basePath.resolve(FILE_NAME + i + FILE_EXT);
         }
 
         boolean hasTombstones = !sstables.isEmpty() && MemoryAccess.getLongAtOffset(sstables.get(0), 16) == 1;
@@ -66,20 +70,14 @@ class Storage implements Closeable {
     }
 
     // it is supposed that entries can not be changed externally during this method call
-    static void save(
-        Config config,
-        Storage previousState,
-        Collection<Entry<MemorySegment>> entries) throws IOException
+    static void save(Config config, Storage previousState, Collection<Entry<MemorySegment>> entries) throws IOException
     {
         int nextSSTableIndex = previousState.sstables.size();
         Path sstablePath = config.basePath().resolve(FILE_NAME + nextSSTableIndex + FILE_EXT);
         save(entries::iterator, sstablePath);
     }
 
-    private static void save(
-        Data entries,
-        Path sstablePath
-    ) throws IOException
+    private static void save(Data entries, Path sstablePath) throws IOException
     {
 
         Path sstableTmpPath = sstablePath.resolveSibling(sstablePath.getFileName().toString() + FILE_EXT_TMP);
@@ -91,8 +89,7 @@ class Storage implements Closeable {
             long size = 0;
             long entriesCount = 0;
             boolean hasTombstone = false;
-            for (Iterator<Entry<MemorySegment>> iterator = entries.iterator(); iterator.hasNext(); ) {
-                Entry<MemorySegment> entry = iterator.next();
+            for (Entry<MemorySegment> entry : entries) {
                 size += getSize(entry);
                 if (entry.isTombstone()) {
                     hasTombstone = true;
@@ -102,18 +99,12 @@ class Storage implements Closeable {
 
             long dataStart = INDEX_HEADER_SIZE + INDEX_RECORD_SIZE * entriesCount;
 
-            MemorySegment nextSSTable = MemorySegment.mapFile(
-                sstableTmpPath,
-                0,
-                dataStart + size,
-                FileChannel.MapMode.READ_WRITE,
-                writeScope
-            );
+            MemorySegment nextSSTable =
+                MemorySegment.mapFile(sstableTmpPath, 0, dataStart + size, FileChannel.MapMode.READ_WRITE, writeScope);
 
             long index = 0;
             long offset = dataStart;
-            for (Iterator<Entry<MemorySegment>> iterator = entries.iterator(); iterator.hasNext(); ) {
-                Entry<MemorySegment> entry = iterator.next();
+            for (Entry<MemorySegment> entry : entries) {
                 MemoryAccess.setLongAtOffset(nextSSTable, INDEX_HEADER_SIZE + index * INDEX_RECORD_SIZE, offset);
 
                 offset += writeRecord(nextSSTable, offset, entry.key());
@@ -132,34 +123,8 @@ class Storage implements Closeable {
         Files.move(sstableTmpPath, sstablePath, StandardCopyOption.ATOMIC_MOVE);
     }
 
-    private static long getSize(Entry<MemorySegment> entry) {
-        if (entry.value() == null) {
-            return Long.BYTES + entry.key().byteSize() + Long.BYTES;
-        } else {
-            return Long.BYTES + entry.value().byteSize() + entry.key().byteSize() + Long.BYTES;
-        }
-    }
-
     public static long getSizeOnDisk(Entry<MemorySegment> entry) {
         return getSize(entry) + INDEX_RECORD_SIZE;
-    }
-
-    private static long writeRecord(MemorySegment nextSSTable, long offset, MemorySegment record) {
-        if (record == null) {
-            MemoryAccess.setLongAtOffset(nextSSTable, offset, -1);
-            return Long.BYTES;
-        }
-        long recordSize = record.byteSize();
-        MemoryAccess.setLongAtOffset(nextSSTable, offset, recordSize);
-        nextSSTable.asSlice(offset + Long.BYTES, recordSize).copyFrom(record);
-        return Long.BYTES + recordSize;
-    }
-
-    @SuppressWarnings("DuplicateThrows")
-    private static MemorySegment mapForRead(ResourceScope scope, Path file) throws NoSuchFileException, IOException {
-        long size = Files.size(file);
-
-        return MemorySegment.mapFile(file, 0, size, FileChannel.MapMode.READ_ONLY, scope);
     }
 
     public static void compact(Config config, Data data) throws IOException {
@@ -178,12 +143,6 @@ class Storage implements Closeable {
 
         Files.move(compactedFile, config.basePath().resolve(FILE_NAME + 0 + FILE_EXT), StandardCopyOption.ATOMIC_MOVE);
     }
-
-    // supposed to have fresh files first
-
-    private final ResourceScope scope;
-    private final ArrayList<MemorySegment> sstables;
-    private final boolean hasTombstones;
 
     private Storage(ResourceScope scope, ArrayList<MemorySegment> sstables, boolean hasTombstones) {
         this.scope = scope;
@@ -208,7 +167,6 @@ class Storage implements Closeable {
         }
         long recordsCount = MemoryAccess.getLongAtOffset(sstable, 8);
         if (key == null) {
-            // fixme
             return recordsCount;
         }
 
@@ -241,10 +199,8 @@ class Storage implements Closeable {
             long keySize = MemoryAccess.getLongAtOffset(sstable, offset);
             long valueOffset = offset + Long.BYTES + keySize;
             long valueSize = MemoryAccess.getLongAtOffset(sstable, valueOffset);
-            return new BaseEntry<>(
-                sstable.asSlice(offset + Long.BYTES, keySize),
-                valueSize == -1 ? null : sstable.asSlice(valueOffset + Long.BYTES, valueSize)
-            );
+            return new BaseEntry<>(sstable.asSlice(offset + Long.BYTES, keySize),
+                valueSize == -1 ? null : sstable.asSlice(valueOffset + Long.BYTES, valueSize));
         } catch (IllegalStateException e) {
             throw checkForClose(e);
         }
@@ -311,16 +267,12 @@ class Storage implements Closeable {
     @Override
     public void close() throws IOException {
         while (scope.isAlive()) {
-            try {
-                scope.close();
-                return;
-            } catch (IllegalStateException ignored) {
-            }
+            scope.close();
         }
     }
 
     public void maybeClose() {
-
+        // no operations
     }
 
     public boolean isClosed() {
@@ -337,7 +289,7 @@ class Storage implements Closeable {
         return !hasTombstones;
     }
 
-    public interface Data {
-        Iterator<Entry<MemorySegment>> iterator() throws IOException;
+    public interface Data extends Iterable<Entry<MemorySegment>> {
+        Iterator<Entry<MemorySegment>> iterator();
     }
 }
