+++ src/main/java/ok/dht/test/monakhov/database/TooManyFlushesInBgException.java
@@ -1,4 +1,3 @@
 package ok.dht.test.monakhov.database;
 
-public class TooManyFlushesInBgException extends RuntimeException {
-}
+public class TooManyFlushesInBgException extends RuntimeException {}
